pyproject.toml
[project]
name = "aiocua"
description = "A Simple Python CUA (Computer Use Agent) Solution"
version = "0.0.0a0"
requires-python = ">=3.10"
readme = "README.md"
license-files = ["LICENSE"]
dependencies = ["pillow"]

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["aiocua*"]
